+ reconstruct exports and tests
