+++ src/helper/mod.rs
@@ -1,6 +1,5 @@
 pub mod callback;
 pub mod parse_identity;
-pub mod parse_permission;
 pub mod parse_recipient;
 pub mod secret_buf;
 pub mod set_owner_group;
