+++ src/parser/mod.rs
@@ -1,0 +1,5 @@
+mod permission;
+mod template;
+
+pub use permission::parse_octal_str;
+pub use template::extract_all_hashes;
